Move ActsTrackingGeometry struct to trackbase
