+++ offline/packages/trackreco/Makefile.am
@@ -24,7 +24,6 @@ AM_LDFLAGS = \
 pkginclude_HEADERS = \
   ActsTransformations.h \
   ActsTrack.h \
-  ActsTrackingGeometry.h \
   ActsSurfaceMaps.h \
   ActsEvaluator.h \
   AssocInfoContainer.h \
