+++ offline/packages/trackreco/PHActsTrackProjection.h
@@ -6,7 +6,7 @@
 #include <trackbase_historic/SvtxTrack.h>
 
 #include "ActsTrack.h"
-#include "ActsTrackingGeometry.h"
+#include <trackbase/ActsTrackingGeometry.h>
 
 #include <Acts/Utilities/Definitions.hpp>
 #include <Acts/Propagator/Propagator.hpp>
