+++ offline/packages/trackreco/PHActsTracks.h
@@ -15,7 +15,7 @@
 #include <ActsExamples/EventData/TrkrClusterSourceLink.hpp>
 
 #include "ActsTrack.h"
-#include "ActsTrackingGeometry.h"
+#include <trackbase/ActsTrackingGeometry.h>
 
 #include <boost/bimap.hpp>
